requirements.txt
fastapi==0.115.8
matplotlib==3.10.0
numpy==2.2.3
optuna==4.2.1
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.5.1+cu124
torchvision==0.20.1+cu124
